Reorganize sql models - avoid circular imports
