+++ qcfractal/tests/test_sqlalchemy.py
@@ -12,9 +12,9 @@ from sqlalchemy.orm import joinedload
 
 import qcfractal.interface as ptl
 from qcfractal.services.services import TorsionDriveService
-from qcfractal.storage_sockets.sql_models import (KVStoreORM, MoleculeORM, OptimizationHistory,
-                                                  OptimizationProcedureORM, ResultORM, ServiceQueueORM, TaskQueueORM,
-                                                  TorsionDriveProcedureORM, Trajectory)
+from qcfractal.storage_sockets import (KVStoreORM, MoleculeORM, OptimizationHistory,
+                                       OptimizationProcedureORM, ResultORM, ServiceQueueORM,
+                                       TaskQueueORM, TorsionDriveProcedureORM, Trajectory)
 from qcfractal.testing import sqlalchemy_socket_fixture as storage_socket
